+++ Docs/Whats_New.md
@@ -25,7 +25,8 @@ Released TBD
 - Allow `MVK_CONFIG_SYNCHRONOUS_QUEUE_SUBMITS` build setting to be overridden.
 - Fix memory leaks of system classes during `VkInstance` and `VkQueue` creation.
 - Fix memory leaks when compiling shaders and pipelines without default OS autorelease pool.
-- Reduce memory usage by adjusting default memory allocs for many MVKVectorInline uses.
+- Reduce memory usage by adjusting default memory allocs for many `MVKVectorInline` uses and 
+  replacing use of `MVKVectorDefault` with `std::vector` in descriptor set bindings.